This is the more in keeping with maintainer preferences commit of first support for most MySQL types. It needs more unit tests, but if they approach won't cut it, I'd rather not waste time writing them. Have ignored Set and Enum types because they can't really be implemented cleanly at the moment. Some of the behavior with Timestamp default/on update values is also likely unintuitive. I very much prefer using in-db functionality for on-create and on-update values, but doing it cleanly will be easier when there's support for commiting to db with only fields that have been assigned values.
Depends on https://github.com/1602/jugglingdb/pull/217 .

I'd really like to see cleaner and clearer support for custom/additional types
in juggling-mysql.

If this commit/merge is not acceptable I'd love an outline of what needs
to be done for something similar to happen. If you don't intend to have clean
support for additional types, also please say so.